% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa2_df.R
\name{pa2_df}
\alias{pa2_df}
\title{Transform API response from list to data frame}
\usage{
pa2_df(response_l, long = FALSE)
}
\arguments{
\item{response_l}{A list based on a Plausible API v2 response, typically
retrieved with \code{\link[=pa2_get]{pa2_get()}}.}

\item{long}{Logical, defaults to FALSE. If FALSE, the default, the value for
each metric is returned in its own column. If TRUE, the data frame is
returned in the long format.}
}
\value{
A data frame.
}
\description{
Transform API response from list to data frame
}
\examples{
\dontrun{
pa2_get(
  date_range = "7d",
  metrics = c("visits", "pageviews"),
  dimensions = c("time:day"),
  pagination = list(limit = 10)
) |>
  pa2_df(long = TRUE)


pa2_get(
  date_range = "7d",
  metrics = c("visits", "pageviews"),
  dimensions = c("time:day"),
  pagination = list(limit = 10)
) |>
  pa2_df(long = FALSE)
}
}
