% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa2_get_metrics_by_day_and_page.R
\name{pa2_get_metrics_by_day_and_page}
\alias{pa2_get_metrics_by_day_and_page}
\title{Retrieve metrics by day and page}
\usage{
pa2_get_metrics_by_day_and_page(
  metrics = c("visits"),
  start_date = Sys.Date() - 8,
  end_date = Sys.Date() - 1,
  long = FALSE,
  filters = NULL,
  order_by = NULL,
  include = NULL,
  pagination = NULL,
  site_id = NULL,
  cache = TRUE,
  cache_connection = NULL,
  only_cached = FALSE,
  wait = 0.1
)
}
\arguments{
\item{metrics}{Defaults to "visits". See
\href{https://plausible.io/docs/stats-api#metrics}{official
documentation on metrics in API version 2} for more details.}

\item{start_date}{Earliest day to include in the output. Defaults to 8 days
ago in order to include the last full week of data. Date is expected in the
format \code{YYYY-MM-DD}, either as character or Date.}

\item{end_date}{Most recent day to include in the output. Defaults to
yesterday in order to include the last full week of data. Date is expected
in the format \code{YYYY-MM-DD}, either as character or Date.}

\item{long}{Logical, defaults to FALSE. If FALSE, the default, the value for
each metric is returned in its own column. If TRUE, the data frame is
returned in the long format.}

\item{filters}{Optional, defaults to \code{NULL}. If given, it must be a list of
three (operator, dimension, clauses) or four (operator, dimension, clauses,
modifiers) elements. See examples. For details,
see the \href{https://plausible.io/docs/stats-api#filters-}{API
documentation on filtering}.}

\item{order_by}{Optional, defaults to \code{NULL}. See
\href{https://plausible.io/docs/stats-api#order_by-}{official
documentation on order in API version 2} for more details.}

\item{include}{Optional, defaults to \code{NULL}, if given must be a named list
(see examples). Additional options for the query as to what data to
include. See the
\href{https://plausible.io/docs/stats-api#include-}{\code{include} section of
the official documentation} for details.}

\item{pagination}{Optional, defaults to \code{NULL}, if given must be a named list
(see examples). Implicitly API default to \code{pagination = list(limit = 10000, offset = 0)}. Define number of results in the results, and change the
offset component to paginate. See
\href{https://plausible.io/docs/stats-api#pagination}{the official
documentation for more details}.}

\item{site_id}{Corresponds to the domain of your website.}

\item{cache}{Defaults to \code{TRUE}. If \code{TRUE}, caches data in a local SQLite
database, stored under a folder named as the website in the current working
directory. The SQLite database is named after the chosen properties, hence
there shouldn't be problems in caching data for different websites or
different combinations of properties.}

\item{cache_connection}{An active database connection, defaults to \code{NULL}
(internally, defaults to a \code{duckdb} database stored in a folder with the
same name as the \code{site_id}).}

\item{only_cached}{Defaults to \code{FALSE}. If \code{TRUE}, only data cached locally will
be retrieved.}

\item{wait}{Numeric, defaults to 0.1. As this function is likely to make a
high number of requests to the API, a small pause is added between each
request to reduce load on the servers.  description}
}
\value{
A data frame with four columns when \code{long} is set to \code{TRUE}
(\code{date}, \code{metric}, \code{value}), or as wide as needed when \code{long} is set to
\code{FALSE}, depending on the number of requested metrics.
}
\description{
Results are cached locally. Caching is filters-aware: data are returned from
cache, only if exactly the same filters are included (or none are present).
}
\examples{
\dontrun{
  pa2_get_metrics_by_day(metrics = c("visits", "visitors", "pageviews"))
}
}
