% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_get_aggregate.R
\name{pa_get_aggregate}
\alias{pa_get_aggregate}
\title{Get aggregated stats for a given website}
\usage{
pa_get_aggregate(period = "30d", metrics = "visits")
}
\arguments{
\item{period}{Period to include in the analysis, defaults to "30d". See
examples as well as the the official documentation for valid values:
\url{https://plausible.io/docs/stats-api#time-dimensions}}

\item{metrics}{Defaults to "visits". See
\href{https://plausible.io/docs/stats-api-v1#metrics}{official
documentation on metrics of API version 1} for more details.}
}
\value{
A data frame.
}
\description{
Get aggregated stats for a given website
}
\examples{
\dontrun{
  pa_get_aggregate(period = "6mo")
}
}
