% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_get_breakdown.R
\name{pa_get_breakdown}
\alias{pa_get_breakdown}
\title{Get a breakdown of stats by a given property}
\usage{
pa_get_breakdown(period = "30d", property, limit = 100)
}
\arguments{
\item{period}{Period to include in the analysis, defaults to "30d". See examples as well as the the official documentation for valid values: https://plausible.io/docs/stats-api#time-periods}

\item{property}{Property to break down the stats by. For a full list of available properties, see the official documentation: https://plausible.io/docs/stats-api#properties}

\item{limit}{Limit the number of results. Defaults to 100.}
}
\value{
A data frame.
}
\description{
Get a breakdown of stats by a given property
}
\examples{
\dontrun{
pa_get_breakdown(period = "30d", property = "event:page")
}
}
