% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_get_aggregate.R
\name{pa_get_page_aggregate}
\alias{pa_get_page_aggregate}
\title{Get aggregated stats for a given page}
\usage{
pa_get_page_aggregate(page, period = "30d", limit = 100)
}
\arguments{
\item{page}{url of relevant page, typically stripped of the top-level domain.
Home page is typically just "/".}

\item{period}{Period to include in the analysis, defaults to "30d". See
examples as well as the the
\href{https://plausible.io/docs/stats-api#time-periods}{official
documentation for valid time periods values}.}

\item{limit}{Limit the number of results. Defaults to 100.}
}
\value{
A data frame.
}
\description{
Get aggregated stats for a given page
}
\examples{
\dontrun{
  pa_get_page_aggregate(period = "6mo", page = "/")
}
}
