% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibler-package.R
\docType{package}
\name{plausibler-package}
\alias{plausibler}
\alias{plausibler-package}
\title{plausibler: Access Plausible Analytics API}
\description{
Access Plausible Analytics API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giocomai.github.io/plausibler/}
}

}
\author{
\strong{Maintainer}: Giorgio Comai \email{g@giorgiocomai.eu} (\href{https://orcid.org/0000-0002-0515-9542}{ORCID})

Authors:
\itemize{
  \item Giorgio Comai \email{g@giorgiocomai.eu} (\href{https://orcid.org/0000-0002-0515-9542}{ORCID})
}

}
\keyword{internal}
