% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa2_get.R
\name{pa2_get}
\alias{pa2_get}
\title{Gets data from Plausible Analytics API (v2)}
\usage{
pa2_get(
  date_range = "30d",
  metrics = "visits",
  dimensions = NULL,
  filters = NULL,
  order_by = NULL,
  include = NULL,
  pagination = NULL,
  site_id = NULL
)
}
\arguments{
\item{filters}{Optional, defaults to NULL. If given, it must be given in the
form "visit:browser==Firefox;visit:country==FR", or as a named vector (see
examples). Use ";" to separate multiple filtering criteria. For details,
see the \href{API documentation on
filtering}{https://plausible.io/docs/stats-api#filters-} for reference.}
}
\value{
A named list, with three elements: \code{results}, \code{meta}, and \code{query}.
}
\description{
For details, consult the \href{official
documentation}{https://plausible.io/docs/stats-api}.
}
\examples{
\dontrun{
pa2_get(
    date_range = "30d",
    metrics = "visits",
)
}
}
