% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa2_get_metrics_by_day.R
\name{pa2_get_metrics_by_day}
\alias{pa2_get_metrics_by_day}
\title{Retrieve metrics by day}
\usage{
pa2_get_metrics_by_day(
  metrics = c("visits"),
  start_date = Sys.Date() - 8,
  end_date = Sys.Date() - 1,
  long = FALSE,
  filters = NULL,
  order_by = NULL,
  include = NULL,
  pagination = NULL,
  site_id = NULL,
  cache = TRUE,
  cache_connection = NULL,
  only_cached = FALSE,
  wait = 0.1
)
}
\arguments{
\item{metrics}{Defaults to "visits". See
\href{https://plausible.io/docs/stats-api#metrics}{official
documentation on metrics in API version 2} for more details.}

\item{long}{Logical, defaults to FALSE. If FALSE, the default, the value for
each metric is returned in its own column. If TRUE, the data frame is
returned in the long format.}

\item{filters}{Optional, defaults to \code{NULL}. If given, it must be a list of
three (operator, dimension, clauses) or four (operator, dimension, clauses,
modifiers) elements. See examples. For details,
see the \href{https://plausible.io/docs/stats-api#filters-}{API
documentation on filtering}.}

\item{order_by}{Optional, defaults to \code{NULL}. See
\href{https://plausible.io/docs/stats-api#order_by-}{official
documentation on order in API version 2} for more details.}

\item{include}{Optional, defaults to \code{NULL}, if given must be a named list
(see examples). Additional options for the query as to what data to
include. See the
\href{https://plausible.io/docs/stats-api#include-}{\code{include} section of
the official documentation} for details.}

\item{pagination}{Optional, defaults to \code{NULL}, if given must be a named list
(see examples). Implicitly API default to \code{pagination = list(limit = 10000, offset = 0)}. Define number of results in the results, and change the
offset component to paginate. See
\href{https://plausible.io/docs/stats-api#pagination-}{the official
documentation for more details}.}

\item{site_id}{Corresponds to the domain of your website.}
}
\value{
A data frame with three columns when \code{long} is set to \code{TRUE}
(\code{date}, \code{metric}, \code{value}), or as wide as needed when \code{long} is set to
\code{FALSE}, depending on the number of requested metrics.
}
\description{
Results are cached locally. Caching is filters-aware: data are returned from
cache, only if exactly the same filters are included (or none are present).
}
\examples{
\dontrun{
pa2_get_metrics_by_day(metrics = c("visits", "visitors", "pageviews"))
}
}
