% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_create_handler.R
\name{pa_create_handler}
\alias{pa_create_handler}
\title{Creates a curl handler object with API keys (used internally)}
\usage{
pa_create_handler()
}
\value{
A curl handler object
}
\description{
Creates a curl handler object with API keys (used internally)
}
\examples{
pa_set(
  base_url = "https://plausible.io/",
  site_id = "example.com",
  key = "actual_key_here"
)

pa_create_handler()
}
