% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_get_breakdown.R
\name{pa_get_breakdown}
\alias{pa_get_breakdown}
\title{Get a breakdown of stats by a given property}
\usage{
pa_get_breakdown(period = "30d", property, limit = 100, filters = NULL)
}
\arguments{
\item{period}{Period to include in the analysis, defaults to "30d". See
examples as well as the the official documentation for valid values:
https://plausible.io/docs/stats-api#time-periods}

\item{property}{Property to break down the stats by. For a full list of
available properties, see the official documentation:
https://plausible.io/docs/stats-api#properties}

\item{limit}{Limit the number of results. Maximum value is 1000. Defaults to
100. If you want to get more than 1000 results, you can make multiple
requests and paginate the results by specifying the page parameter (e.g.
make the same request with page=1, then page=2, etc)}

\item{filters}{Optional, defaults to NULL. If given, it must be given in the
form "visit:browser==Firefox;visit:country==FR", or as a named vector (see
examples). Use ";" to separate multiple filtering criteria. For details,
see the \href{API documentation on
filtering}{https://plausible.io/docs/stats-api#filtering} for reference.}
}
\value{
A data frame.
}
\description{
Get a breakdown of stats by a given property
}
\examples{
\dontrun{
pa_get_breakdown(period = "30d", property = "event:page")

## With filters, e.g. to see all referrers to the given url:

pa_get_breakdown(period = "30d",
                 property = "visit:referrer",
                 filters = list(`event:page` = "/berlin/"))
}
}
