% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_set.R
\name{pa_get_settings}
\alias{pa_get_settings}
\title{Retrieves session settings}
\usage{
pa_get_settings(base_url = NULL, site_id = NULL, key = NULL)
}
\arguments{
\item{base_url}{The base URL of the Plausible instance. Set to
https://plausible.io/ - or to your own domain for self-hosted Plausible.}

\item{site_id}{Corresponds to the domain of your website.}

\item{key}{A character string used for authentication. Can be retrieved from
the user settings in Plausible Analytics.}
}
\value{
A list (invisibly).
}
\description{
It allows to override specific settings, without changing environment
variables.
}
\examples{
pa_set(
  base_url = "https://plausible.io/",
  site_id = "example.com",
  key = "actual_key_here"
)

pa_get_settings()

## It allows to override specific settings, without changing environment variables

pa_get_settings(site_id = "notexample.com")
}
