% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_get_timeseries.R
\name{pa_get_timeseries}
\alias{pa_get_timeseries}
\title{Get time series}
\usage{
pa_get_timeseries(period)
}
\arguments{
\item{period}{Period to include in the analysis. See the relevant part of the official documentation for valid values}
}
\value{
A data frame.
}
\description{
Get time series
}
\examples{
pa_get_timeseries(period = "6mo")
pa_get_timeseries(period = "30d")
pa_get_timeseries(period = "custom&date=2021-06-01,2021-06-07")

}
